Move the preparation of the webm namespace to webm.py
